requirements.txt
astroid==2.9.3
attrs==21.4.0
bandit==1.7.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.4
colorlog==4.8.0
cryptography==36.0.1
dataclasses==0.8
decorator==4.4.2
Deprecated==1.2.13
dulwich==0.20.32
feedparser==6.0.8
flake8==4.0.1
flake8-polyfill==1.0.2
Flask==2.0.3
future==0.18.2
gitdb==4.0.9
GitPython==3.1.18
graal==0.2.9
grimoirelab-toolkit==0.2.0
html-to-json==2.0.0
idna==3.3
importlib-metadata==4.2.0
ipython-genutils==0.2.0
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==3.2.0
jupyter-core==4.9.2
lazy-object-proxy==1.7.1
lizard==1.16.6
mando==0.6.4
MarkupSafe==2.0.1
mccabe==0.6.1
nbformat==5.1.3
networkx==2.5.1
pbr==5.8.1
perceval==0.17.16
platformdirs==2.4.0
plotly==4.14.3
progress==1.4
pycodestyle==2.8.0
pycparser==2.21
pydot==1.4.2
pyflakes==2.4.0
PyGithub==1.55
PyJWT==2.3.0
pylint==2.12.2
PyNaCl==1.5.0
pyparsing==3.0.7
pyrsistent==0.18.0
python-dateutil==2.8.2
PyYAML==6.0
radon==4.0.0
requests==2.27.1
retrying==1.3.3
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.1
stevedore==3.5.0
tabulate==0.8.9
toml==0.10.2
traitlets==4.3.3
typed-ast==1.5.2
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
wily==1.20.0
wrapt==1.13.3
zipp==3.6.0
